+++ packages/icons/__tests__/android/icons.test.js
@@ -2,15 +2,6 @@
 
 import shared from "../shared";
 
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "android";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.android || obj.default);
-  return reactNative;
-});
-
 describe("Icons test on android: ", () => {
   shared();
 });