+++ packages/pagination/__tests__/android/pagination.native.test.js
@@ -1,0 +1,4 @@
+import tests from "../test-helper";
+import Pagination from "../../pagination";
+
+describe("Pagination on Android", tests(Pagination));