+++ packages/pagination/__tests__/test-wrapper-helper.js
@@ -1,10 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import { Text } from "react-native";
-import Enzyme, { shallow } from "enzyme";
-import React16Adapter from "enzyme-adapter-react-16";
-
-Enzyme.configure({ adapter: new React16Adapter() });
+import { shallow } from "enzyme";
 
 export default withPageState => () => {
   it("renders inner component with page 1", () => {