+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.17"></a>
+## [0.6.17](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.6.16...@times-components/article-label@0.6.17) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-label
+
 <a name="0.6.16"></a>
 ## [0.6.16](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.6.15...@times-components/article-label@0.6.16) (2018-01-19)